tui: Extract action prompt menu formatting
The interactive prompt module handled readline setup, prompt input, and
action menu wrapping. The wrapping code owns terminal width, visible ANSI
length, and grouped menu section layout.

Keeping that formatter in tui.prompts made the prompt module broader than
input handling and made menu rendering details harder to name.

This commit moves action menu section formatting into
tui.action_prompt_menu and has tui.prompts call that module when rendering
grouped action sections.

Subsequent commits will add architecture coverage for the action prompt
menu boundary.
